app/requirements.txt
djangorestframework
drf_spectacular
djangorestframework-simplejwt
drf-extensions
mysqlclient